Review comments
* Use Deque instead of Stack
* Simplify the OSGi test by removing unnecessary duplication from AdapterTest
